requirements.txt
mysql-connector-python
pymongo
neo4j
redis
python-dotenv
